requirements.txt
aiogram>=3.0.0
aio-mysql
python-dotenv